Optimizations and fixes

+++ app/ore/OreConfig.scala
@@ -80,7 +80,7 @@ final class OreConfig @Inject()(config: Configuration) {
       val maxNameLen: Int               = raw.get[Int]("max-name-len")
       val maxPages: Int                 = raw.get[Int]("max-pages")
       val maxChannels: Int              = raw.get[Int]("max-channels")
-      val initLoad: Int                 = raw.get[Int]("init-load")
+      val initLoad: Long                = raw.get[Long]("init-load")
       val initVersionLoad: Int          = raw.get[Int]("init-version-load")
       val maxDescLen: Int               = raw.get[Int]("max-desc-len")
       val fileValidate: Boolean         = raw.get[Boolean]("file-validate")
@@ -91,11 +91,11 @@ final class OreConfig @Inject()(config: Configuration) {
     }
 
     object users extends ConfigCategory {
-      val raw: Configuration   = ore.raw.get[Configuration]("users")
-      val starsPerPage: Int    = raw.get[Int]("stars-per-page")
-      val maxTaglineLen: Int   = raw.get[Int]("max-tagline-len")
-      val authorPageSize: Long = raw.get[Long]("author-page-size")
-      val projectPageSize: Int = raw.get[Int]("project-page-size")
+      val raw: Configuration    = ore.raw.get[Configuration]("users")
+      val starsPerPage: Int     = raw.get[Int]("stars-per-page")
+      val maxTaglineLen: Int    = raw.get[Int]("max-tagline-len")
+      val authorPageSize: Long  = raw.get[Long]("author-page-size")
+      val projectPageSize: Long = raw.get[Long]("project-page-size")
     }
 
     object orgs extends ConfigCategory {
